Sprint1/backend/requirements.txt
Authlib==1.3.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
cryptography==42.0.5
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonify==0.5
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
requests-file==2.0.0
setuptools==69.2.0
tldextract==5.1.2
urllib3==2.2.1
Werkzeug==3.0.2
zope.event==5.0
zope.interface==6.2
